+++ nginx-consul/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.1.7
+
+* Make consul-template retry more often
+
+---
+
 0.1.6
 
 * Update consul configuration to new version